Replaced hrefs in Header with NavLinks. App.js refactored to include BrowserRouter and migrate existing code to a new homepage folder called  Home.
