+++ package.json
@@ -10,7 +10,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-particles": "^2.9.3",
-    "react-scripts": "5.0.1",
+    "react-router-hash-link": "^2.4.3",
+    "react-scripts": "^5.0.1",
     "tsparticles": "^2.9.3",
     "tsparticles-engine": "^2.9.3",
     "tsparticles-preset-triangles": "^2.9.3",
