+++ src/Home/CountDownTimer.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import './App.css';
+import './Home.css';
 
 function CountdownTimer() {
     const [countdown, setCountdown] = useState(Date.parse('2023-04-16T00:00:00') - Date.now());
